fixes for camera, keyboard for players
