+++ public_html/js/views/GameModules/gameInit.js
@@ -23,64 +23,24 @@ define(function (require) {
             gameObjects.renderer = new THREE.WebGLRenderer();
             this.container = $('#game-canvas');
 
-            gameObjects.firstCharacter = new Character.init({color: 0xff0000}, {x: 5, z: 7});
+            gameObjects.firstCharacter = new Character.init({color: 0xff0000}, {x: 31, z: 31});
             
             
             var lol = new  Character.init({color: 0x00FF00}, {x: 4, z: 9});
 
             gameObjects.scene.add(lol.mesh);
             gameObjects.scene.add(gameObjects.firstCharacter.mesh);
+            gameObjects.firstCharacter.setControls('top');
 
             World.init();
             gameObjects.scene.add(World.mesh);
+            gameObjects.firstCharacter.setFocus(gameObjects.firstCharacter.mesh, -950);
 
-            this.setAspect();
-            this.container.prepend(gameObjects.renderer.domElement);
-            
-            // Set the camera to look at our user's character
-            this.setFocus(gameObjects.firstCharacter.mesh);
-            this.setControls();
-
-        },
-        setControls: function () {
-            var controls = {
-                left: false,
-                up: false,
-                right: false,
-                down: false
-            };
-            function makeControls(status, keyCode) {
-                var pressed = false;
-                if (status === 'pressed') {
-                    pressed = true;
-                }
-                switch (String.fromCharCode(keyCode)) {
-                    case 'A':
-                        controls.left = pressed;
-                        break;
-                    case 'W':
-                        controls.up = pressed;
-                        break;
-                    case 'D':
-                        controls.right = pressed;
-                        break;
-                    case 'S':
-                        controls.down = pressed;
-                        break;
-                }
-                gameObjects.firstCharacter.setDirection(controls);
-            }
-            jQuery(document).keydown(function (e) {
-               makeControls('pressed', e.keyCode);
-                e.preventDefault();
-            });
-            jQuery(document).keyup(function (e) {
-                makeControls('none', e.keyCode);
-                e.preventDefault();
-            });
             jQuery(window).resize(function () {
                 BasicScene.setAspect();
             });
+            this.setAspect();
+            this.container.prepend(gameObjects.renderer.domElement);
         },
 
         setAspect: function () {
@@ -90,13 +50,10 @@ define(function (require) {
             gameObjects.camera.aspect = w / h;
             gameObjects.camera.updateProjectionMatrix();
         },
-        setFocus: function (object) {
-            gameObjects.camera.position.set(object.position.x, object.position.y + 400, object.position.z - 950);
-            gameObjects.camera.lookAt(object.position);
-        },
+
         frame: function () {
             gameObjects.firstCharacter.motion();
-            this.setFocus(gameObjects.firstCharacter.mesh);
+            gameObjects.firstCharacter.setFocus(gameObjects.firstCharacter.mesh , -950);
             gameObjects.renderer.render(gameObjects.scene, gameObjects.camera);
         }
     };
