+++ public_html/js/views/GameModules/worldBuilder.js
@@ -56,6 +56,7 @@ define(function (require) {
             gameObjects.addObjectToWorld(this.worldObjects.destructible_crate, new THREE.CubeGeometry(64, 64, 64), 1, 12, 17);
             gameObjects.addObjectToWorld(this.worldObjects.indestructible_crate, new THREE.CubeGeometry(64, 64, 64),2, 0, 0);
             gameObjects.addObjectToWorld(this.worldObjects.destructible_crate, new THREE.CubeGeometry(64, 64, 64), 3, 2, 2);
+            gameObjects.addObjectToWorld(this.worldObjects.destructible_crate, new THREE.CubeGeometry(64, 64, 64), 4, 29, 31);
             
             
         },
